+++ R/data_rescale.R
@@ -1,12 +1,13 @@
-#' Rescale Variables to a New Range
+#' @title Rescale Variables to a New Range
+#' @name rescale
 #'
+#' @description
 #' Rescale variables to a new range. Can also be used to reverse-score variables
 #' (change the keying/scoring direction), or to expand a range.
 #'
 #' @inheritParams categorize
 #' @inheritParams extract_column_names
 #' @inheritParams standardize.data.frame
-#'
 #' @param to Numeric vector of length 2 giving the new range that the variable
 #'   will have after rescaling. To reverse-score a variable, the range should
 #'   be given with the maximum value first. See examples.
@@ -26,6 +27,11 @@
 #' @param ... Arguments passed to or from other methods.
 #'
 #' @inheritSection center Selection of variables - the `select` argument
+#'
+#' @seealso See [makepredictcall.dw_transformer()] for use in model formulas.
+#' @family transform utilities
+#'
+#' @return A rescaled object.
 #'
 #' @examples
 #' rescale(c(0, 1, 5, -5, -2))
@@ -62,14 +68,6 @@
 #' # Specify list of multipliers
 #' d <- data.frame(x = 5:15, y = 5:15)
 #' rescale(d, multiply = list(x = 1.1, y = 0.5))
-#'
-#' @inherit data_rename
-#'
-#' @return A rescaled object.
-#'
-#' @seealso See [makepredictcall.dw_transformer()] for use in model formulas.
-#' @family transform utilities
-#'
 #' @export
 rescale <- function(x, ...) {
   UseMethod("rescale")
@@ -84,7 +82,6 @@ change_scale <- function(x, ...) {
 }
 
 
-
 #' @export
 rescale.default <- function(x, verbose = TRUE, ...) {
   if (isTRUE(verbose)) {
@@ -94,7 +91,6 @@ rescale.default <- function(x, verbose = TRUE, ...) {
   }
   x
 }
-
 
 
 #' @rdname rescale
@@ -233,7 +229,6 @@ rescale.grouped_df <- function(x,
   attributes(x) <- utils::modifyList(info, attributes(x))
   x
 }
-
 
 
 #' @rdname rescale